Move administration services into Administration feature folders #388
